Changed  available physical memory to return the available physical ram, instead of total physical ram.
Fixed it so TotalAvailableMemoryBytes updates.

White Space Changes.

Fixed it so TotalAvailableMemoryBytes updates.

Use the heap limit if set, isntead of available memory.

Various fixes.

Fixes so we use the heap limit if set.
