+++ src/mono/mono/metadata/mono-perfcounters.c
@@ -416,77 +416,6 @@ predef_cleanup (ImplVtable *vtable)
 	perfctr_unlock ();
 }
 
-static guint64
-mono_determine_physical_ram_available_size (void)
-{
-#if defined (TARGET_WIN32)
-	MEMORYSTATUSEX memstat;
-
-	memstat.dwLength = sizeof (memstat);
-	GlobalMemoryStatusEx (&memstat);
-	return (guint64)memstat.ullAvailPhys;
-
-#elif defined (__NetBSD__)
-	struct vmtotal vm_total;
-	guint64 page_size;
-	int mib[2];
-	size_t len;
-
-	mib[0] = CTL_VM;
-	mib[1] = VM_METER;
-
-	len = sizeof (vm_total);
-	sysctl (mib, 2, &vm_total, &len, NULL, 0);
-
-	mib[0] = CTL_HW;
-	mib[1] = HW_PAGESIZE;
-
-	len = sizeof (page_size);
-	sysctl (mib, 2, &page_size, &len, NULL, 0);
-
-	return ((guint64) vm_total.t_free * page_size) / 1024;
-#elif defined (__APPLE__)
-	mach_msg_type_number_t count = HOST_VM_INFO_COUNT;
-	mach_port_t host = mach_host_self();
-	vm_size_t page_size;
-	vm_statistics_data_t vmstat;
-	kern_return_t ret;
-	do {
-		ret = host_statistics(host, HOST_VM_INFO, (host_info_t)&vmstat, &count);
-	} while (ret == KERN_ABORTED);
-
-	if (ret != KERN_SUCCESS) {
-		g_warning ("Mono was unable to retrieve memory usage!");
-		return 0;
-	}
-
-	host_page_size(host, &page_size);
-	return (guint64) vmstat.free_count * page_size;
-
-#elif defined (HAVE_SYSCONF)
-	guint64 page_size = 0, num_pages = 0;
-
-	/* sysconf works on most *NIX operating systems, if your system doesn't have it or if it
-	 * reports invalid values, please add your OS specific code below. */
-#ifdef _SC_PAGESIZE
-	page_size = (guint64)sysconf (_SC_PAGESIZE);
-#endif
-
-#ifdef _SC_AVPHYS_PAGES
-	num_pages = (guint64)sysconf (_SC_AVPHYS_PAGES);
-#endif
-
-	if (!page_size || !num_pages) {
-		g_warning ("Your operating system's sysconf (3) function doesn't correctly report physical memory size!");
-		return 0;
-	}
-
-	return page_size * num_pages;
-#else
-	return 0;
-#endif
-}
-
 void
 mono_perfcounters_init (void)
 {
