+++ src/mono/mono/sgen/sgen-gc.h
@@ -748,7 +748,6 @@ typedef struct _SgenGCInfo {
 	guint64 heap_size_bytes;
 	guint64 high_memory_load_threshold_bytes;
 	guint64 memory_load_bytes;
-	guint64 total_available_memory_bytes;
 } SgenGCInfo;
 
 extern SgenGCInfo sgen_gc_info;
